XG7Menus BETA 2.0

+++ src/main/java/com/xg7plugins/xg7menus/api/menus/builders/item/BookItemBuilder.java
@@ -1,12 +1,18 @@
 package com.xg7plugins.xg7menus.api.menus.builders.item;
 
+import com.xg7plugins.xg7menus.api.XG7Menus;
 import com.xg7plugins.xg7menus.api.menus.MenuException;
 import com.xg7plugins.xg7menus.api.menus.builders.BaseItemBuilder;
 import com.xg7plugins.xg7menus.api.utils.NMSUtil;
 import com.xg7plugins.xg7menus.api.utils.Text;
 import com.xg7plugins.xg7menus.api.utils.XSeries.XMaterial;
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.Unpooled;
 import lombok.SneakyThrows;
 import net.md_5.bungee.api.chat.BaseComponent;
+import net.md_5.bungee.api.chat.TextComponent;
+import net.md_5.bungee.chat.TextComponentSerializer;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.BookMeta;
@@ -14,12 +20,14 @@ import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.nio.ByteBuffer;
 
 public class BookItemBuilder extends BaseItemBuilder<BookItemBuilder> {
     public BookItemBuilder() {
-        super(XMaterial.WRITABLE_BOOK.parseItem());
+        super(XMaterial.WRITTEN_BOOK.parseItem());
     }
     public BookItemBuilder(ItemStack book) {
         super(book);
@@ -27,7 +35,7 @@ public class BookItemBuilder extends BaseItemBuilder<BookItemBuilder> {
 
     @Contract("_ -> new")
     public static @NotNull BookItemBuilder from(@NotNull ItemStack book) {
-        if (!book.getType().equals(XMaterial.WRITABLE_BOOK.parseMaterial())) throw new MenuException("This item isn't a writable book!");
+        if (!book.getType().equals(XMaterial.WRITTEN_BOOK.parseMaterial())) throw new MenuException("This item isn't a writable book!");
         return new BookItemBuilder(book);
     }
     @Contract(" -> new")
@@ -53,43 +61,44 @@ public class BookItemBuilder extends BaseItemBuilder<BookItemBuilder> {
         return this;
     }
     public BookItemBuilder addPage(BaseComponent[] components) {
-        BookMeta meta = (BookMeta) this.itemStack.getItemMeta();
-        meta.spigot().addPage(components);
-        super.meta(meta);
+
+        try {
+            BookMeta meta = (BookMeta) this.itemStack.getItemMeta();
+            meta.spigot().addPage(components);
+            super.meta(meta);
+            return this;
+        } catch (Exception ignored) {}
+
         return this;
     }
 
     @SneakyThrows
     public void openBook(Player player) {
+
+        if (Integer.parseInt(Bukkit.getServer().getVersion().split("\\.")[1].replace(")", "")) > 13) {
+            player.openBook(this.itemStack);
+            return;
+        }
+
         int slot = player.getInventory().getHeldItemSlot();
         ItemStack old = player.getInventory().getItem(slot);
         player.getInventory().setItem(slot, this.itemStack);
 
+        ByteBuf buf = Unpooled.buffer(256);
+        buf.setByte(0, 0);
+        buf.writerIndex(1);
+
+        Object packet = NMSUtil.getNMSClass("PacketPlayOutCustomPayload")
+                .getConstructor(
+                        String.class,
+                        NMSUtil.getNMSClass("PacketDataSerializer")
+                ).newInstance("MC|BOpen", NMSUtil.getNMSClass("PacketDataSerializer").getConstructor(ByteBuf.class).newInstance(buf));
         Class<?> craftPlayerClass = NMSUtil.getCraftBukkitClass("entity.CraftPlayer");
-        Class<?> entityPlayerClass = NMSUtil.getNMSClass("EntityPlayer");
+        Object cPlayer = NMSUtil.getCraftBukkitClass("entity.CraftPlayer").cast(player);
+        Object craftPlayerHandle = craftPlayerClass.getMethod("getHandle").invoke(cPlayer);
+        Object playerConnection = craftPlayerHandle.getClass().getField("playerConnection").get(craftPlayerHandle);
+        playerConnection.getClass().getMethod("sendPacket", NMSUtil.getNMSClass("Packet")).invoke(playerConnection, packet);
 
-        Class<?> nmsItemStackClass = NMSUtil.getNMSClass("ItemStack");
-
-        Method getHandle = craftPlayerClass.getMethod("getHandle");
-
-        Object entityPlayer = getHandle.invoke(player);
-
-        Method asNMSCopy = NMSUtil.getCraftBukkitClass("inventory.CraftItemStack").getMethod("asNMSCopy", ItemStack.class);
-        Object nmsItemStack = asNMSCopy.invoke(null, this.itemStack);
-
-        Class<?> packetPlayOutOpenBookClass = NMSUtil.getNMSClass("PacketPlayOutOpenBook");
-
-        Class<?> enumHandClass = NMSUtil.getNMSClass("EnumHand");
-
-        Object mainHand = enumHandClass.getField("MAIN_HAND").get(null);
-
-        Constructor<?> packetConstructor = packetPlayOutOpenBookClass.getConstructor(enumHandClass);
-        Object packet = packetConstructor.newInstance(mainHand);
-
-        Method sendPacket = entityPlayerClass.getField("playerConnection").getType().getMethod("sendPacket", NMSUtil.getNMSClass("Packet"));
-
-        sendPacket.invoke(entityPlayerClass.getField("playerConnection").get(entityPlayer), packet);
-
-        player.getInventory().setItem(slot,old);
+        player.getInventory().setItem(slot, old);
     }
 }
